refactor mac for multiple bindings
